pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ospo_stats"
description = "Open-source project statistics"
version = "v0.0.0"
authors = [{name = "Jason Lo", email = "lcmjlo@gmail.com"}]
requires-python = ">=3.11"
dependencies = ["tenacity", "requests", "python-dotenv", "pandas", "altair"]

[tool.hatch.build.targets.wheel]
include = ["ospo_stats/*.py"]
